c317: extract BenchmarksHidsag tab cluster (HidsagCrossPreprocessingStability + CrossPreprocessingMatrix + HidsagPreprocessing + PreprocessingSubsetCard + PolicyBars + HidsagBenchmarks + HidsagSubsetCard) to own file (#441 P0 2.2)
Benchmarks.tsx 935 LOC -> 333 LOC (-602 LOC). Benchmarks chunk 18.56 KB -> 8.10 KB (-10.5 KB).
New chunk: BenchmarksHidsag-D42dA8hR.js 10.96 KB.
Lifted HIDSAG_SUBSETS const to pages/benchmarks/shared.ts.
Aliased HidsagCrossPreprocessingStability type to avoid local-function name clash.
typecheck clean, 44/44 tests pass.
